requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
anyio==4.9.0
    # via httpx
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.2.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
python-telegram-bot==22.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
typing-extensions==4.13.2
    # via anyio
